requirements.txt
pandas
yfinance
matplotlib
PyYAML
requests
python-dotenv
